# backend/src/main/java/com/scotlandyard/dto/MrXLogEntryDTO.java
package com.scotlandyard.dto;

import com.scotlandyard.model.TicketType;

public class MrXLogEntryDTO {
    private int round;
    private int leg;
    private TicketType ticketUsed;
    private Integer nodeId;

    public int getRound()                { return round; }
    public void setRound(int round)      { this.round = round; }

    public int getLeg()                  { return leg; }
    public void setLeg(int leg)          { this.leg = leg; }

    public TicketType getTicketUsed()            { return ticketUsed; }
    public void setTicketUsed(TicketType t)      { this.ticketUsed = t; }

    public Integer getNodeId()           { return nodeId; }
    public void setNodeId(Integer nodeId){ this.nodeId = nodeId; }
}